# api_service/openapi/public-api-v1.yaml
openapi: 3.1.0
info:
  title: SuggestArr Public API
  version: v1
  description: Public API for integrating with a SuggestArr installation.
...
paths:
...
  /api/v1/webhooks:
    get:
      operationId: publicWebhooks
      tags: [Webhooks]
      summary: List outbound webhooks
      responses:
        <<: *errorResponses
        '200': {description: Webhooks, with secrets omitted}
        '403': {$ref: '#/components/responses/Forbidden'}
    post:
      operationId: publicCreateWebhook
      tags: [Webhooks]
      summary: Create an outbound webhook
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/WebhookCreate'}
      responses:
        <<: *errorResponses
        '201': {description: Webhook created, with secret omitted}
        '403': {$ref: '#/components/responses/Forbidden'}
  /api/v1/webhooks/{webhook_id}:
    delete:
      operationId: publicDeleteWebhook
      tags: [Webhooks]
      summary: Delete an outbound webhook
      parameters:
        - {name: webhook_id, in: path, required: true, schema: {type: string}}
      responses:
        <<: *errorResponses
        '204': {description: Webhook deleted}
        '403': {$ref: '#/components/responses/Forbidden'}
        '404': {description: Webhook not found}
...
  /api/v1/me:
    get:
      operationId: publicMe
      tags: [Identity]
      summary: Current API identity
      description: Shows the user and the authentication method used for this public request.
      responses:
        '200': {description: Current public API identity, content: {application/json: {schema: {$ref: '#/components/schemas/IdentityResponse'}}}}
        <<: *errorResponses
  /api/v1/jobs:
    get:
      operationId: publicJobs
      tags: [Jobs]
      summary: List visible jobs
      description: Administrators see all jobs; other users see only their own jobs.
      parameters:
        - {name: page, in: query, schema: {type: integer, minimum: 1}}
        - {name: per_page, in: query, schema: {type: integer, minimum: 1, maximum: 100}}
        - {name: enabled, in: query, schema: {type: boolean}, description: Filter enabled or disabled jobs.}
        - {name: job_type, in: query, schema: {type: string, enum: [discover, recommendation, trakt_recommendations]}}
      responses:
        <<: *errorResponses
        '200': {description: Visible jobs, content: {application/json: {schema: {$ref: '#/components/schemas/JobListResponse'}}}}
  /api/v1/jobs/{job_id}:
    get:
      operationId: publicJob
      tags: [Jobs]
      summary: Get a visible job
      parameters:
        - {name: job_id, in: path, required: true, schema: {type: integer}}
      responses:
        <<: *errorResponses
        '200': {description: Visible job, content: {application/json: {schema: {$ref: '#/components/schemas/JobResponse'}}}}
        '404': {description: Job not found}
  /api/v1/jobs/{job_id}/preview:
    post:
      operationId: publicJobPreview
      tags: [Jobs]
      summary: Preview a job
      description: Runs the existing dry-run pipeline without submitting requests or writing execution history.
      parameters:
        - {name: job_id, in: path, required: true, schema: {type: integer}}
      responses:
        <<: *errorResponses
        '200': {description: Dry-run result, content: {application/json: {schema: {$ref: '#/components/schemas/PreviewResponse'}}}}
  /api/v1/jobs/{job_id}/runs:
    post:
      operationId: publicStartRun
      tags: [Jobs]
      summary: Start an asynchronous job run
      description: Returns a queued run resource. Poll the Location response header until its status is final.
      parameters:
        - {name: job_id, in: path, required: true, schema: {type: integer}}
      responses:
        <<: *errorResponses
        '202': {description: Queued run, headers: {Location: {schema: {type: string}, description: URL of the queued run}}, content: {application/json: {schema: {$ref: '#/components/schemas/RunResponse'}}}}
  /api/v1/runs:
    get:
      operationId: publicRuns
      tags: [Runs]
      summary: List visible runs
      parameters:
        - {name: page, in: query, schema: {type: integer, minimum: 1}}
        - {name: per_page, in: query, schema: {type: integer, minimum: 1, maximum: 100}}
        - {name: job_id, in: query, schema: {type: integer, minimum: 1}}
        - {name: status, in: query, schema: {type: string, enum: [queued, running, completed, failed, skipped]}}
      responses:
        <<: *errorResponses
        '200': {description: Visible runs, content: {application/json: {schema: {$ref: '#/components/schemas/RunListResponse'}}}}
  /api/v1/runs/{run_id}:
    get:
      operationId: publicRun
      tags: [Runs]
      summary: Get a visible run
      parameters:
        - {name: run_id, in: path, required: true, schema: {type: integer}}
      responses:
        <<: *errorResponses
        '200': {description: Visible run, content: {application/json: {schema: {$ref: '#/components/schemas/RunResponse'}}}}
  /api/v1/suggestions:
    get:
      operationId: publicSuggestions
      tags: [Suggestions]
      summary: List visible suggestions
      parameters:
        - {name: page, in: query, schema: {type: integer, minimum: 1}}
        - {name: per_page, in: query, schema: {type: integer, minimum: 1, maximum: 100}}
        - {name: status, in: query, schema: {type: string, default: awaiting_approval, enum: [all, awaiting_approval, queued, submitting, submitted, rejected, failed, blacklisted]}}
        - {name: media_type, in: query, schema: {type: string, default: all, enum: [all, movie, tv]}}
        - {name: search, in: query, schema: {type: string, maxLength: 100}, description: Match a title or TMDb ID.}
        - {name: user_id, in: query, schema: {type: string}, description: Filter by media-service user ID when permitted by the installation visibility policy.}
      responses:
        <<: *errorResponses
        '200': {description: Visible suggestions, content: {application/json: {schema: {$ref: '#/components/schemas/SuggestionListResponse'}}}}
  /api/v1/suggestions/actions:
    post:
      operationId: publicSuggestionAction
      tags: [Suggestions]
      summary: Apply a suggestion action
      description: A caller can approve, reject, blacklist, retry, or request again for up to 100 accessible suggestions.
      requestBody:
        required: true
        description: Specify one action and from 1 to 100 accessible suggestion IDs.
        content:
          application/json:
            schema: {$ref: '#/components/schemas/SuggestionAction'}
            example:
              action: approve
              ids: [123]
              remove_blacklist: false
      responses:
        <<: *errorResponses
        '200': {description: Action result, content: {application/json: {schema: {$ref: '#/components/schemas/SuggestionActionResponse'}}}}
  /api/v1/requests:
    get:
      operationId: publicRequests
      tags: [Requests]
      summary: List visible requests
      parameters:
        - {name: page, in: query, schema: {type: integer, minimum: 1}}
        - {name: per_page, in: query, schema: {type: integer, minimum: 1, maximum: 100}}
        - {name: sort, in: query, schema: {type: string, default: date_desc, enum: [date_desc, date_asc, title_asc, title_desc, rating_desc, rating_asc]}}
        - {name: source, in: query, schema: {type: string}, description: Exact source-origin filter.}
        - {name: user_id, in: query, schema: {type: string}, description: Filter by media-service user ID when permitted by the installation visibility policy.}
      responses:
        <<: *errorResponses
        '200': {description: Visible requests, content: {application/json: {schema: {$ref: '#/components/schemas/RequestListResponse'}}}}
  /api/v1/requests/stats:
    get:
      operationId: publicRequestStats
      tags: [Requests]
      summary: Get visible request statistics
      parameters:
        - {name: user_id, in: query, schema: {type: string}, description: Filter by media-service user ID when permitted by the installation visibility policy.}
      responses:
        <<: *errorResponses
        '200': {description: Visible request statistics, content: {application/json: {schema: {$ref: '#/components/schemas/RequestStatsResponse'}}}}
  /api/v1/installation/stats:
    get:
      operationId: publicInstallationStats
      tags: [Installation]
      summary: Get installation statistics
      description: Administrator-only aggregate counters for requests, jobs, executions, suggestions, and the Seer delivery queue. This endpoint intentionally exposes counts only, never configuration or credentials.
      responses:
        <<: *errorResponses
        '200': {description: Installation-wide operational statistics, content: {application/json: {schema: {$ref: '#/components/schemas/InstallationStatsResponse'}}}}
        '403': {$ref: '#/components/responses/Forbidden'}
components:
  securitySchemes:
    bearerAuth: {type: http, scheme: bearer, bearerFormat: JWT}
    apiKeyAuth: {type: apiKey, in: header, name: X-API-Key}
  responses:
...
    Forbidden:
      description: Authenticated identity is not an administrator
      content:
        application/json:
          schema: {$ref: '#/components/schemas/Error'}
...
  schemas:
...
    IdentityResponse:
      type: object
      required: [data]
      properties:
        data:
          type: object
          required: [id, username, role, permissions, authentication]
          properties:
            id: {type: [integer, string]}
            username: {type: string}
            role: {type: string, enum: [admin, user]}
            permissions:
              type: object
              required: [can_manage_ai]
              properties: {can_manage_ai: {type: boolean}}
            authentication:
              type: object
              required: [method, api_key_id, api_key_name]
              properties:
                method: {type: string, enum: [jwt, api_key, local_bypass, disabled]}
                api_key_id: {type: [integer, 'null']}
                api_key_name: {type: [string, 'null']}
    JobResponse:
      type: object
      properties: {data: {$ref: '#/components/schemas/Job'}}
    JobListResponse:
      type: object
      properties: {data: {type: array, items: {$ref: '#/components/schemas/Job'}}, meta: {$ref: '#/components/schemas/Meta'}}
    RunResponse:
      type: object
      properties: {data: {$ref: '#/components/schemas/Run'}}
    RunListResponse:
      type: object
      properties: {data: {type: array, items: {$ref: '#/components/schemas/Run'}}, meta: {$ref: '#/components/schemas/Meta'}}
    PreviewResponse:
      type: object
      properties: {data: {type: object, properties: {job_id: {type: integer}, items_count: {type: integer}, items: {type: array, items: {type: object, additionalProperties: true}}}}}
...
    SuggestionListResponse:
      type: object
      required: [data, meta]
      properties:
        data: {type: array, items: {$ref: '#/components/schemas/Suggestion'}}
        meta: {$ref: '#/components/schemas/Meta'}
    RequestListResponse:
      type: object
      required: [data, meta]
      properties:
        data: {type: array, items: {$ref: '#/components/schemas/Request'}}
        meta: {$ref: '#/components/schemas/Meta'}
    SuggestionActionResponse:
      type: object
      properties: {data: {type: object, properties: {action: {type: string}, updated: {type: integer}}}}
    RequestStatsResponse:
      type: object
      properties: {data: {$ref: '#/components/schemas/RequestStatistics'}}
    InstallationStatsResponse:
      type: object
      properties:
        data:
          type: object
          properties:
            requests: {$ref: '#/components/schemas/RequestStatistics'}
            jobs: {$ref: '#/components/schemas/JobStatistics'}
            runs: {$ref: '#/components/schemas/ExecutionStatistics'}
            suggestions: {$ref: '#/components/schemas/SuggestionStatistics'}
            queue: {$ref: '#/components/schemas/QueueStatistics'}
...
    WebhookCreate:
      type: object
      required: [name, url, secret, events]
      properties:
        name: {type: string, minLength: 1, maxLength: 100}
        url: {type: string, format: uri, maxLength: 2048}
        secret: {type: string, minLength: 16, writeOnly: true}
        events:
          type: array
          minItems: 1
          items: {type: string, enum: [suggestion.created, suggestion.awaiting_approval, suggestion.approved, suggestion.rejected, request.submitted, request.failed, run.failed]}
        enabled: {type: boolean, default: true}
        allow_private: {type: boolean, default: false, description: Allow private-network delivery destinations.}
...
    SuggestionAction:
      type: object
      required: [action, ids]
      properties:
        action: {type: string, enum: [approve, reject, blacklist, retry, request_again], description: Action to apply to every supplied suggestion.}
        ids:
          type: array
          minItems: 1
          maxItems: 100
          uniqueItems: true
          items: {type: integer}
          description: Suggestion IDs to update.
        remove_blacklist: {type: boolean, default: false, description: Valid only with request_again; removes a matching blacklist entry.}
security:
  - bearerAuth: []
  - apiKeyAuth: []